MM_30482 : Re-organizing operators
