+++ app/database/admin/data_operator/operators/utils.ts
@@ -1,0 +1,22 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import DatabaseManager from '@database/admin/database_manager';
+import {DatabaseType} from '@typings/database/enums';
+
+jest.mock('@database/admin/database_manager');
+
+export const createConnection = async (databaseName: string) => {
+    const serverUrl = 'https://appv2.mattermost.com';
+    const database = await DatabaseManager.createDatabaseConnection({
+        shouldAddToDefaultDatabase: true,
+        configs: {
+            actionsEnabled: true,
+            dbName: databaseName,
+            dbType: DatabaseType.SERVER,
+            serverUrl,
+        },
+    });
+
+    return database;
+};
